+++ src/main/java/generics/ethereum/EthereumWallet.java
@@ -4,7 +4,7 @@ import generics.Wallet;
 
 public class EthereumWallet extends Wallet<Ethereum> {
 
-    public EthereumWallet(Ethereum currency, int currencyQuantity) {
+    public EthereumWallet(Ethereum currency, double currencyQuantity) {
         super(currency, currencyQuantity);
     }
