+++ src/main/java/generics/standard/StandardCurrencyHolder.java
@@ -4,5 +4,5 @@ import generics.CurrencyHolder;
 
 public interface StandardCurrencyHolder<T extends StandardCurrency> extends CurrencyHolder<T> {
 
-    void transferCurrencyToBankAccount(int currencyAmountToTransfer, T currency);
+    void transferCurrencyToBankAccount(double currencyAmountToTransfer, T currency);
 }
