+++ src/main/java/generics/standard/StandardWallet.java
@@ -4,7 +4,7 @@ import generics.Wallet;
 
 public abstract class StandardWallet<T extends StandardCurrency> extends Wallet<T> implements StandardCurrencyHolder<T> {
 
-    public StandardWallet(T currency, int currencyQuantity) {
+    public StandardWallet(T currency, double currencyQuantity) {
         super(currency, currencyQuantity);
     }
